refactor(replsession): extract observation and summary helpers
